Greatly beef up testing framework, implement and document rest of basic methods
The test framework now has a library module (t::lib) which arranges to
set up and tear down a test Maildir with embedded .mu, so that we
always test in isolation.  You can run the tests against your own Maildir
with
  $ env MAILDIR=$HOME/Maildir make test
.. if you're crazy and go for that sort of thing.

Also importing a sample Maildir tree with two messages in it, to
ease testing.