Add support for (:update (...)) messages from server, fix add test.
We now support an update_callback member which optionally can have
a coderef to a callback to invoke whenever an (:update (...)) message
is seen from the server; it sends these (undocumented!?) messages
when e.g. an 'add' command adds a message, in addition to the result
of the 'add' command itself.

Still working this stuff out...